Implements dequeueAll reusing idea from Buffer filterInPlace
The algorithm follow Buffer.filterInPlace, but instead of keeping the elements matching predicate, we skip it and store it in new elems collection, in the queue keeping only the ones that does not match the condition.

I also removed the work around on test/files/run/QueueTest.scala and move it into JUnit.

Fixes scala/collection-strawman#530